+++ Network/API/Telegram/Bot/Object/Update/Message/Origin.hs
@@ -1,9 +1,9 @@
 module Network.API.Telegram.Bot.Object.Update.Message.Origin (Origin (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
-import "aeson" Data.Aeson.Types (Object, Parser, Value)
+import "aeson" Data.Aeson.Types (Object, Parser, Value (Object))
 import "base" Control.Applicative (Applicative ((<*>)))
-import "base" Control.Monad (Monad ((>>=)), fail)
+import "base" Control.Monad (Monad ((>>=)))
 import "base" Data.Eq (Eq ((==)))
 import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
@@ -12,12 +12,12 @@ import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
 import Network.API.Telegram.Bot.Object.Sender (Sender)
+import Network.API.Telegram.Bot.Object.Update.Moving.Group (Group)
 import Network.API.Telegram.Bot.Property.Identifiable (Identifiable (Identificator, ident))
 
 data Origin
 	= Private Int64 Sender
-	| Group Int64 Text Sender
-	| Supergroup Int64 Text Sender
+	| Group Group Sender
 	| Channel Int64 Text
 	deriving Show
 
@@ -27,8 +27,7 @@ instance Eq Origin where
 instance Identifiable Origin where
 	type Identificator Origin = Int64
 	ident (Private i _) = i
-	ident (Group i _ _) = i
-	ident (Supergroup i _ _) = i
+	ident (Group g _) = ident g
 	ident (Channel i _) = i
 
 instance FromJSON Origin where
@@ -37,7 +36,5 @@ instance FromJSON Origin where
 		chat :: Object -> Value -> Parser Origin
 		chat v = withObject "Origin" $ \c -> c .: "type" >>= \case
 			("private" :: Text) -> Private <$> c .: "id" <*> v .: "from"
-			("group" :: Text) -> Group <$> c .: "id" <*> c .: "title" <*> v .: "from"
-			("supergroup" :: Text) -> Supergroup <$> c .: "id" <*> c .: "title" <*> v .: "from"
 			("channel" :: Text) -> Channel <$> c .: "id" <*> c .: "title"
-			_ -> fail "Type of chat is not defined"
+			_ -> parseJSON (Object v)