+++ Network/API/Telegram/Bot/Object.hs
@@ -2,4 +2,5 @@ module Network.API.Telegram.Bot.Object (module Exports) where
 
 import Network.API.Telegram.Bot.Object.Update as Exports
 import Network.API.Telegram.Bot.Object.Sender as Exports
+import Network.API.Telegram.Bot.Object.Name as Exports
 import Network.API.Telegram.Bot.Object.Member as Exports