+++ Network/API/Telegram/Bot/Object/Sender.hs
@@ -1,4 +1,4 @@
-module Network.API.Telegram.Bot.Object.Sender (Sender (..), nickname, firstname, lastname) where
+module Network.API.Telegram.Bot.Object.Sender (Sender (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), Object, withObject, (.:), (.:?))
 import "aeson" Data.Aeson.Types (Parser)
@@ -11,14 +11,15 @@ import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
 import "base" Data.Maybe (Maybe)
 import "base" Text.Show (Show)
-import "lens" Control.Lens (Lens')
 import "text" Data.Text (Text)
 
+import Network.API.Telegram.Bot.Object.Name (Name, First, Last, Nick)
+import Network.API.Telegram.Bot.Property.Accessible (Accessible (access))
 import Network.API.Telegram.Bot.Property.Identifiable (Identifiable (Identificator, ident))
 
 data Sender
-	= Bot Int (Maybe Text) Text (Maybe Text) (Maybe Text)
-	| Human Int (Maybe Text) Text (Maybe Text) (Maybe Text)
+	= Bot Int (Maybe (Nick Name)) (First Name) (Maybe (Last Name)) (Maybe Text)
+	| Human Int (Maybe (Nick Name)) (First Name) (Maybe (Last Name)) (Maybe Text)
 	deriving Show
 
 instance Eq Sender where
@@ -31,7 +32,19 @@ instance Identifiable Sender where
 	ident (Bot i _ _ _ _) = i
 	ident (Human i _ _ _ _) = i
 
-type Whom = Int -> Maybe Text -> Text -> Maybe Text -> Maybe Text -> Sender
+instance Accessible (First Name) Sender where
+	access f (Bot uid nn fn ln lng) = (\fn' -> Bot uid nn fn' ln lng) <$> f fn
+	access f (Human uid nn fn ln lng) = (\fn' -> Human uid nn fn' ln lng) <$> f fn
+
+instance Accessible (Maybe (Last Name)) Sender where
+	access f (Bot uid nn fn ln lng) = (\ln' -> Bot uid nn fn ln' lng) <$> f ln
+	access f (Human uid nn fn ln lng) = (\ln' -> Human uid nn fn ln' lng) <$> f ln
+
+instance Accessible (Maybe (Nick Name)) Sender where
+	access f (Bot uid nn fn ln lng) = (\nn' -> Bot uid nn' fn ln lng) <$> f nn
+	access f (Human uid nn fn ln lng) = (\nn' -> Human uid nn' fn ln lng) <$> f nn
+
+type Whom = Int -> Maybe (Nick Name) -> First Name -> Maybe (Last Name) -> Maybe Text -> Sender
 
 instance FromJSON Sender where
 	parseJSON = withObject "Sender" $ \v -> v .: "is_bot"
@@ -41,15 +54,3 @@ instance FromJSON Sender where
 		sender v f = f <$> v .: "id" <*> v .:? "username"
 			<*> v .: "first_name" <*> v .:? "last_name"
 			<*> v .:? "language_code"
-
-nickname :: Lens' Sender (Maybe Text)
-nickname f (Bot uid nn fn ln lng) = (\nn' -> Bot uid nn' fn ln lng) <$> f nn
-nickname f (Human uid nn fn ln lng) = (\nn' -> Human uid nn' fn ln lng) <$> f nn
-
-firstname :: Lens' Sender Text
-firstname f (Bot uid nn fn ln lng) = (\fn' -> Bot uid nn fn' ln lng) <$> f fn
-firstname f (Human uid nn fn ln lng) = (\fn' -> Human uid nn fn' ln lng) <$> f fn
-
-lastname :: Lens' Sender (Maybe Text)
-lastname f (Bot uid nn fn ln lng) = (\ln' -> Bot uid nn fn ln' lng) <$> f ln
-lastname f (Human uid nn fn ln lng) = (\ln' -> Human uid nn fn ln' lng) <$> f ln