+++ CHANGELOG.md
@@ -130,3 +130,4 @@
 # 0.2.6
 * Define `Eq` instance for `Update`, `Message`, `Callback`, `Origin`, `File` and `Size`
 * Start to write documentation step-by-step, show three simple examples
+* Add `Sender` field into `Callback` datatype to understand from whom action goes