+++ CHANGELOG.md
@@ -132,3 +132,4 @@
 * Start to write documentation step-by-step, show three simple examples
 * Add `Sender` field into `Callback` datatype to understand from whom action goes
 * Change endpoint for `Edit`ing `Keyboard` object, remove `Edit` `Text :&: Keyboard` instance
+* Replace `Group` and `Supergroup` constructors by including `Group` field in `Origin` datatype