+++ CHANGELOG.md
@@ -133,3 +133,6 @@
 * Add `Sender` field into `Callback` datatype to understand from whom action goes
 * Change endpoint for `Edit`ing `Keyboard` object, remove `Edit` `Text :&: Keyboard` instance
 * Replace `Group` and `Supergroup` constructors by including `Group` field in `Origin` datatype
+
+# 0.2.7
+* Define `Name` module to make `Accessible` instances for `Nick`/`First`/`Last` names